+++ src/services/nostr.ts
@@ -19,7 +19,14 @@ import {
 import { createEventLoaderForStore } from "applesauce-loaders/loaders";
 import { RelayPool } from "applesauce-relay";
 import { map, takeUntil, timer } from "rxjs";
-import { CACHE_RELAYS, LOOKUP_RELAYS, NOSTR_RELAYS } from "../env.ts";
+import {
+  CACHE_RELAYS,
+  LOOKUP_RELAYS,
+  NOSTR_RELAYS,
+  OUTBOXES_STALE_TIME,
+  PROFILES_STALE_TIME,
+  SERVERS_STALE_TIME,
+} from "../env.ts";
 import logger from "../helpers/debug.ts";
 import { formatAgeFromUnix } from "../helpers/format.ts";
 import { onShutdown } from "../helpers/shutdown.ts";
@@ -32,10 +39,7 @@ import * as cache from "../services/cache.ts";
 
 const log = logger.extend("nostr");
 
-const SITE_MANIFEST_KINDS = [
-  ROOT_SITE_MANIFEST_KIND,
-  NAMED_SITE_MANIFEST_KIND,
-];
+const SITE_MANIFEST_KINDS = [ROOT_SITE_MANIFEST_KIND, NAMED_SITE_MANIFEST_KIND];
 
 export const pool = new RelayPool();
 
@@ -46,7 +50,9 @@ eventStore.filters({ kinds: [BLOSSOM_SERVER_LIST_KIND] }).subscribe((list) => {
   const servers = getBlossomServersFromList(list);
   log(
     `Found ${servers.length} blossom servers for ${list.pubkey}, id=${list.id}, age=${
-      formatAgeFromUnix(list.created_at)
+      formatAgeFromUnix(
+        list.created_at,
+      )
     }\n`,
     servers.join(", "),
   );
@@ -58,7 +64,9 @@ eventStore.filters({ kinds: [kinds.RelayList] }).subscribe((list) => {
     `Found ${
       outboxes.length + inboxes.length
     } relays for ${list.pubkey}, id=${list.id}, age=${
-      formatAgeFromUnix(list.created_at)
+      formatAgeFromUnix(
+        list.created_at,
+      )
     }\n`,
     inboxes.join(", "),
     "\n",
@@ -103,17 +111,19 @@ export async function syncSiteManifests(
   return await new Promise((resolve, reject) => {
     let found = 0;
 
-    pool.request(relays, {
-      kinds: SITE_MANIFEST_KINDS,
-      ...(since === undefined ? {} : { since }),
-    }).subscribe({
-      next: (event) => {
-        const insert = eventStore.add(event);
-        if (insert && insert === event) found++;
-      },
-      error: reject,
-      complete: () => resolve(found),
-    });
+    pool
+      .request(relays, {
+        kinds: SITE_MANIFEST_KINDS,
+        ...(since === undefined ? {} : { since }),
+      })
+      .subscribe({
+        next: (event) => {
+          const insert = eventStore.add(event);
+          if (insert && insert === event) found++;
+        },
+        error: reject,
+        complete: () => resolve(found),
+      });
   });
 }
 
@@ -128,43 +138,55 @@ if (CACHE_RELAYS) {
   });
 }
 
+const profilesChecked = new Map<string, Date>();
+
 /** Gets a user's profile */
 export async function getUserProfile(
   pubkey: string,
   timeout = 5_000,
 ): Promise<ProfileContent | undefined> {
+  const checked = profilesChecked.get(pubkey);
+  const stale = checked &&
+    checked.getTime() > Date.now() - PROFILES_STALE_TIME * 1000;
+
   // check in-memory cache
   const cached = eventStore.getReplaceable(kinds.Metadata, pubkey);
-  if (cached) return getProfileContent(cached);
 
-  // Check deno kv cache
-  const lvl2Cache = await cache.getUserProfile(pubkey);
-  // Return if found, or explicitly null
-  if (lvl2Cache) return lvl2Cache === 404 ? undefined : lvl2Cache;
+  // If results are fresh, return whatever is in the event store
+  if (!stale) return cached && getProfileContent(cached);
 
-  // Using first value because loading profiles is not critical and needs to be fast
+  // Otherwise, fetch the latest profile
   const user = castUser(pubkey, eventStore);
   const profile = await firstValueFrom(
     user.profile$.metadata.pipe(takeUntil(timer(timeout))),
-    { defaultValue: undefined },
+    {
+      defaultValue: undefined,
+    },
   );
 
-  // Set deno kv cache
-  if (profile) cache.setUserProfile(pubkey, profile ?? null).catch(() => {});
+  // Set the last checked time
+  profilesChecked.set(pubkey, new Date());
 
   return profile;
 }
+
+const outboxesChecked = new Map<string, Date>();
 
 /** Gets a user's outbox relays */
 export async function getUserOutboxes(
   pubkey: string,
   timeout = 5_000,
 ): Promise<string[] | undefined> {
+  const checked = outboxesChecked.get(pubkey);
+  const stale = checked &&
+    checked.getTime() > Date.now() - OUTBOXES_STALE_TIME * 1000;
   const cached = eventStore.getReplaceable(kinds.RelayList, pubkey);
-  if (cached) return getOutboxes(cached);
 
+  // If results are fresh, return whatever is in the event store
+  if (!stale) return cached && getOutboxes(cached);
+
+  // Otherwise, fetch the latest outboxes
   const user = castUser(pubkey, eventStore);
-
   const outboxes = await lastValueFrom(
     user.outboxes$.pipe(
       // Add hard timeout
@@ -173,14 +195,28 @@ export async function getUserOutboxes(
     { defaultValue: undefined },
   );
 
+  // Set the last checked time
+  outboxesChecked.set(pubkey, new Date());
+
   return outboxes;
 }
 
+const blossomServersChecked = new Map<string, Date>();
+
 /** Gets a users list of blossom servers */
 export async function getUserBlossomServers(pubkey: string, timeout = 5_000) {
-  const cached = eventStore.getReplaceable(BLOSSOM_SERVER_LIST_KIND, pubkey);
-  if (cached) return getBlossomServersFromList(cached).map((s) => s.toString());
+  const checked = blossomServersChecked.get(pubkey);
+  const stale = checked &&
+    checked.getTime() > Date.now() - SERVERS_STALE_TIME * 1000;
 
+  const cached = eventStore.getReplaceable(BLOSSOM_SERVER_LIST_KIND, pubkey);
+
+  // If results are fresh, return whatever is in the event store
+  if (!stale) {
+    return cached && getBlossomServersFromList(cached).map((s) => s.toString());
+  }
+
+  // Otherwise, fetch the latest blossom servers
   const user = castUser(pubkey, eventStore);
   const servers = await lastValueFrom(
     user.blossomServers$.pipe(
@@ -189,6 +225,10 @@ export async function getUserBlossomServers(pubkey: string, timeout = 5_000) {
     ),
     { defaultValue: undefined },
   );
+
+  // Set the last checked time
+  blossomServersChecked.set(pubkey, new Date());
+
   return servers?.map((server) => server.toString());
 }
 
@@ -201,9 +241,7 @@ export async function getManifest(address: AddressPointer, timeout = 5_000) {
   );
   if (manifest) return manifest;
 
-  log(
-    `Loading manifest ${getReplaceableAddressFromPointer(address)}`,
-  );
+  log(`Loading manifest ${getReplaceableAddressFromPointer(address)}`);
 
   const outboxes = await getUserOutboxes(address.pubkey, timeout);
   const relays = relaySet(outboxes, address.relays);
@@ -237,8 +275,6 @@ export function isMatchingManifestAddress(
       getManifestIdentifier(event) === undefined;
   }
 
-  return (
-    event.kind === NAMED_SITE_MANIFEST_KIND &&
-    getManifestIdentifier(event) === identifier
-  );
+  return event.kind === NAMED_SITE_MANIFEST_KIND &&
+    getManifestIdentifier(event) === identifier;
 }